requirements.txt
requests
pyinotify
sqlalchemy
enum34
scielo.plumber
scieloapi
-e git+git://github.com/scieloorg/packtools.git#egg=packtools
pyramid
pyramid_xmlrpc
WebTest
transaction
zope.sqlalchemy
psycopg2
lxml
factory_boy
paramiko
alembic

